feat(FriendList): 화면 수정, 폴더 구조 수정, 컴포넌트 분리
