+++ src/data/users.ts
@@ -1,0 +1,28 @@
+export const DUMMY_USERS = [
+    { id: '1', nickname: '닉네임', email: 'test@example.com' },
+    { id: '2', nickname: '닉네임', email: 'test@example.com' },
+    { id: '3', nickname: '닉네임', email: 'test@example.com' },
+    { id: '4', nickname: '닉네임', email: 'test@example.com' },
+    { id: '5', nickname: '닉네임', email: 'test@example.com' },
+    { id: '6', nickname: '닉네임', email: 'test@example.com' },
+    { id: '7', nickname: '닉네임', email: 'test@example.com' },
+    { id: '8', nickname: '닉네임', email: 'test@example.com' },
+    { id: '9', nickname: '닉네임', email: 'test@example.com' },
+    { id: '10', nickname: '닉네임', email: 'test@example.com' },
+    { id: '11', nickname: '닉네임', email: 'test@example.com' },
+    { id: '12', nickname: '닉네임', email: 'test@example.com' },
+    { id: '13', nickname: '닉네임', email: 'test@example.com' },
+    { id: '14', nickname: '닉네임', email: 'test@example.com' },
+    { id: '15', nickname: '닉네임', email: 'test@example.com' },
+    { id: '16', nickname: '닉네임16', email: 'test@example.com' },
+    { id: '17', nickname: '스크롤1', email: 'scroll1@example.com' },
+    { id: '18', nickname: '스크롤2', email: 'scroll2@example.com' },
+    { id: '19', nickname: '스크롤3', email: 'scroll3@example.com' },
+    { id: '20', nickname: '스크롤4', email: 'scroll4@example.com' },
+    { id: '21', nickname: '스크롤5', email: 'scroll5@example.com' },
+    { id: '22', nickname: '스크롤6', email: 'scroll6@example.com' },
+    { id: '23', nickname: '스크롤7', email: 'scroll7@example.com' },
+    { id: '24', nickname: '스크롤8', email: 'scroll8@example.com' },
+    { id: '25', nickname: '스크롤9', email: 'scroll9@example.com' },
+    { id: '26', nickname: '스크롤10', email: 'scroll10@example.com' },
+  ];
